refactor: move `Mentor.jsx` contents into `mentors.jsx`
